[flang][Frontend] Move LangOptions from Frontend to Common
The information in LangOptions is not tied to any frontend code,
and could be used by any other component.
