+++ flang/lib/Common/CMakeLists.txt
@@ -40,6 +40,7 @@ add_flang_library(FortranCommon
   Fortran-features.cpp
   default-kinds.cpp
   idioms.cpp
+  LangOptions.cpp
   Version.cpp
   ${version_inc}
